refactor(result): change result server component
